requirements.txt
attrs==19.1.0
awscli==1.6.2
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.7.0.1
bcdoc==0.12.2
bleach==3.1.0
boto==2.33.0
botocore==0.73.0
certifi==2015.4.28
colorama==0.2.5
configparser==3.7.4
decorator==4.0.9
defusedxml==0.6.0
docutils==0.14
entrypoints==0.3
enum34==1.1.6
functools32==3.2.3.post2
futures==3.2.0
genson==0.1.0
GraphLab-Create==2.1
GraphLab-Create-License==2.1
ipaddress==1.0.22
ipykernel==4.10.0
ipython==5.8.0
ipython-genutils==0.2.0
Jinja2==2.10.1
jmespath==0.5.0
jsonschema==2.5.1
jupyter-client==5.2.4
jupyter-core==4.4.0
MarkupSafe==1.1.1
mistune==0.8.4
multipledispatch==0.6.0
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
pandocfilters==1.4.2
pathlib2==2.3.3
pexpect==4.7.0
pickleshare==0.7.5
prettytable==0.7.2
prometheus-client==0.6.0
prompt-toolkit==1.0.15
psclient==2.0
ptyprocess==0.6.0
pyasn1==0.4.5
Pygments==2.4.0
pyrsistent==0.14.11
python-dateutil==2.8.0
pyzmq==18.0.0
requests==2.9.1
rsa==3.1.2
scandir==1.10.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
sseclient==0.0.8
terminado==0.8.2
testpath==0.4.2
tornado==4.3
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
